Extract PrePurchaseUpgradesErrorView to own file
